Moved DTOs to a dedicated folder in `Borea.Storage.Index` for folder cleanup. Fixed `InstallInfoDto`
