+++ src/Borea.Storage/Index/Dtos/AnyOfDependencyDto.cs
@@ -1,18 +1,15 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Borea.Storage.Index;
+namespace Borea.Storage.Index.Dtos;
 
-public sealed class LoaderDto
+public sealed class AnyOfDependencyDto
 {
     [JsonPropertyName("id")]
     public required string Id { get; set; }
 
     [JsonPropertyName("min")]
-    public required string Min { get; set; }
+    public string? Min { get; set; }
 
     [JsonPropertyName("max")]
     public string? Max { get; set; }
-
-    [JsonPropertyName("source")]
-    public string? Source { get; set; }
 }
