+++ src/Borea.Storage/Index/Dtos/InstallDescriptorDto.cs
@@ -1,14 +1,11 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Borea.Storage.Index;
+namespace Borea.Storage.Index.Dtos;
 
-public sealed class InstallInfoDto
+public sealed class InstallDescriptorDto
 {
     [JsonPropertyName("root")]
     public string? Root { get; set; }
-
-    [JsonPropertyName("derived")]
-    public required bool Derived { get; set; }
 
     [JsonPropertyName("target")]
     public string? Target { get; set; }
