+++ src/Borea.Storage/Index/Dtos/ListingEntryDto.cs
@@ -1,7 +1,7 @@
 ﻿using System.Text.Json;
 using System.Text.Json.Serialization;
 
-namespace Borea.Storage.Index;
+namespace Borea.Storage.Index.Dtos;
 
 public sealed class ListingEntryDto
 {
