+++ src/Borea.Storage/Index/Dtos/LoaderDto.cs
@@ -1,15 +1,18 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Borea.Storage.Index;
+namespace Borea.Storage.Index.Dtos;
 
-public sealed class AnyOfDependencyDto
+public sealed class LoaderDto
 {
     [JsonPropertyName("id")]
     public required string Id { get; set; }
 
     [JsonPropertyName("min")]
-    public string? Min { get; set; }
+    public required string Min { get; set; }
 
     [JsonPropertyName("max")]
     public string? Max { get; set; }
+
+    [JsonPropertyName("source")]
+    public string? Source { get; set; }
 }
